(Defect) 1. invalidate also clear the corresponding bits 2. modify the singletons into pointers 3. rename ca to fp

+++ src/chunk/chunkmodule.h
@@ -15,7 +15,7 @@ namespace cache {
     bool next(Chunk &c);
     bool next(uint64_t &addr, uint8_t *&buf, uint32_t &len);
    protected:
-    uint32_t _addr;
+    uint64_t _addr;
     uint8_t *_buf;
     uint32_t _len;
     uint32_t _chunk_size;
